llvmpipe: Fix overflow for 32 bits available memory computation
And cap to 2 GB on 32 bits.

Fixes https://bugs.freedesktop.org/show_bug.cgi?id=94561

+++ src/gallium/auxiliary/os/os_misc.c
@@ -128,7 +128,7 @@ os_get_total_physical_memory(uint64_t *size)
    const long phys_pages = sysconf(_SC_PHYS_PAGES);
    const long page_size = sysconf(_SC_PAGE_SIZE);
 
-   *size = phys_pages * page_size;
+   *size = (int64_t)phys_pages * (int64_t)page_size;
    return (phys_pages > 0 && page_size > 0);
 #elif defined(PIPE_OS_APPLE) || defined(PIPE_OS_BSD)
    size_t len = sizeof(*size);
